requirements.txt
requests==2.25.*
pathvalidate-2.4.1
beautifulsoup4==4.9.3
lxml==4.6.3